modem-manager/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyserial==3.5
pyaudio==0.2.11
numpy==1.25.2
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
webrtcvad==2.0.10
noisereduce==3.0.0
httpx==0.25.2
asyncio-mqtt==0.16.1
pydantic==2.5.0
python-dotenv==1.0.0
psutil==5.9.6
aiofiles==23.2.1